import libmongoc 1.14.0

+++ src/mongoc/mongoc-iovec.h
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
+#include "mongoc/mongoc-prelude.h"
+
 
 #ifndef MONGOC_IOVEC_H
 #define MONGOC_IOVEC_H
 
 
-#include <bson.h>
+#include <bson/bson.h>
 
 #ifdef _WIN32
 #include <stddef.h>
@@ -32,7 +34,7 @@ BSON_BEGIN_DECLS
 
 #ifdef _WIN32
 typedef struct {
-   u_long iov_len;
+   size_t iov_len;
    char *iov_base;
 } mongoc_iovec_t;
 
@@ -40,10 +42,10 @@ BSON_STATIC_ASSERT2 (sizeof_iovect_t,
                      sizeof (mongoc_iovec_t) == sizeof (WSABUF));
 BSON_STATIC_ASSERT2 (offsetof_iovec_base,
                      offsetof (mongoc_iovec_t, iov_base) ==
-                     offsetof (WSABUF, buf));
+                        offsetof (WSABUF, buf));
 BSON_STATIC_ASSERT2 (offsetof_iovec_len,
                      offsetof (mongoc_iovec_t, iov_len) ==
-                     offsetof (WSABUF, len));
+                        offsetof (WSABUF, len));
 
 #else
 typedef struct iovec mongoc_iovec_t;
